Async type check.
